pyproject.toml
[tool.poetry]
name = "zaechka-bot"
version = "0.1.0"
description = "Telegram bot that loves you!"
authors = ["Artyom Sosedka <sostema@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
python-telegram-bot = "^13.13"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
